(REF) Civi/API/Event - Extract RequestTrait
